app/requirements.txt
flask
flask-sqlalchemy
SQLAlchemy-Searchable
SQLAlchemy-Utils
psycopg2-binary
gunicorn

